backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic==2.6.0
pydantic-settings==2.2.0
python-jose[cryptography]==3.3.0
google-auth==2.28.0
google-generativeai==0.4.1
python-multipart==0.0.9
httpx==0.26.0
